+++ thumblr_platform_interface/lib/src/platform_interface/thumblr_platform_interface.dart
@@ -1,5 +1,3 @@
-import 'dart:ui' as ui;
-
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:thumblr_platform_interface/src/method_channel/method_channel_thumblr.dart';
 
@@ -36,7 +34,7 @@ abstract class ThumblrPlatform extends PlatformInterface {
   }
 
   /// Generates a thumbnail for a given video file.
-  Future<ui.Image> generateThumbnail({
+  Future<Map<String, dynamic>> generateThumbnail({
     required String filePath,
     double position = 0.0,
   }) {